+++ sentry-android-sqlite/src/main/java/io/sentry/android/sqlite/SQLiteSpanManager.kt
@@ -11,7 +11,8 @@ import io.sentry.SpanStatus
 private const val TRACE_ORIGIN = "auto.db.sqlite"
 
 internal class SQLiteSpanManager(
-    private val hub: IHub = HubAdapter.getInstance()
+    private val hub: IHub = HubAdapter.getInstance(),
+    private val databaseName: String? = null
 ) {
     private val stackTraceFactory = SentryStackTraceFactory(hub.options)
 
@@ -46,6 +47,15 @@ internal class SQLiteSpanManager(
                 if (isMainThread) {
                     setData(SpanDataConvention.CALL_STACK_KEY, stackTraceFactory.inAppCallStack)
                 }
+                // if db name is null, then it's an in-memory database as per
+                // https://cs.android.com/androidx/platform/frameworks/support/+/androidx-main:sqlite/sqlite/src/main/java/androidx/sqlite/db/SupportSQLiteOpenHelper.kt;l=38-42
+                if (databaseName != null) {
+                    setData(SpanDataConvention.DB_SYSTEM_KEY, "sqlite")
+                    setData(SpanDataConvention.DB_NAME_KEY, databaseName)
+                } else {
+                    setData(SpanDataConvention.DB_SYSTEM_KEY, "in-memory")
+                }
+
                 finish()
             }
         }