+++ sentry/src/main/java/io/sentry/Sentry.java
@@ -360,7 +360,9 @@ public final class Sentry {
     }
 
     final @NotNull IModulesLoader modulesLoader = options.getModulesLoader();
-    if (modulesLoader instanceof NoOpModulesLoader) {
+    if (!options.isSendModules()) {
+      options.setModulesLoader(NoOpModulesLoader.getInstance());
+    } else if (modulesLoader instanceof NoOpModulesLoader) {
       options.setModulesLoader(
           new CompositeModulesLoader(
               Arrays.asList(