+++ sentry/src/test/java/io/sentry/SentryTest.kt
@@ -6,6 +6,7 @@ import io.sentry.internal.debugmeta.IDebugMetaLoader
 import io.sentry.internal.debugmeta.ResourcesDebugMetaLoader
 import io.sentry.internal.modules.CompositeModulesLoader
 import io.sentry.internal.modules.IModulesLoader
+import io.sentry.internal.modules.NoOpModulesLoader
 import io.sentry.protocol.SdkVersion
 import io.sentry.protocol.SentryId
 import io.sentry.test.ImmediateExecutorService
@@ -32,6 +33,7 @@ import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
+import kotlin.test.assertIs
 import kotlin.test.assertNotEquals
 import kotlin.test.assertNotNull
 import kotlin.test.assertNull
@@ -785,6 +787,18 @@ class SentryTest {
             anyOrNull(),
             anyOrNull()
         )
+    }
+
+    @Test
+    fun `if send modules is false, uses NoOpModulesLoader`() {
+        var sentryOptions: SentryOptions? = null
+        Sentry.init {
+            it.dsn = dsn
+            it.isSendModules = false
+            sentryOptions = it
+        }
+
+        assertIs<NoOpModulesLoader>(sentryOptions?.modulesLoader)
     }
 
     private class InMemoryOptionsObserver : IOptionsObserver {